"Metastatic Spread During Treatment" scatter plot complete
Reduce the Placebo's marker's width in both the the "Tumor Response to 
Treatment" and "Metastatic Spread During Treatment" scatter plots to the 
smallest size to prevent the Placebo's markers from completely covering 
Ketapril's markers.
Replace the "x_axis" variable with a more dynamic value to condense the 
code and adjust for any potential future changes to the data set.