Wrangle the data for the survival rates plot and declare an x-axis
The data is properly formatted to construct the survival rates plot.
All of the scatter plots have the same x-axis values. Assign x_axis to 
the value used as the first plot's, "Tumor Response to Treatment", 
x-axis and use x_axis in all of the plots to condense the code.