Construct a second bar plot and alter the first bar plot as necessary
Construct the second bar plot so the drugs that passed can be separated 
from the drugs that failed.
Remove the width parameter assignment and leave only the width value in 
ax.bar() for both bar plots to prevent the number of bars from doubling.
Set the x-axis positions so the x-axis tick labels can be set outside of 
.bar().
Set the x-axis tick labels outside of .bar() so the labels do not 
overlap.